requirements.txt
pyobjc==2.4
pyobjc-core==2.4
pyobjc-framework-Accounts==2.4
pyobjc-framework-AddressBook==2.4
pyobjc-framework-AppleScriptKit==2.4
pyobjc-framework-AppleScriptObjC==2.4
pyobjc-framework-Automator==2.4
pyobjc-framework-CFNetwork==2.4
pyobjc-framework-CalendarStore==2.4
pyobjc-framework-Cocoa==2.4
pyobjc-framework-Collaboration==2.4
pyobjc-framework-CoreData==2.4
pyobjc-framework-CoreLocation==2.4
pyobjc-framework-CoreText==2.4
pyobjc-framework-DictionaryServices==2.4
pyobjc-framework-EventKit==2.4
pyobjc-framework-ExceptionHandling==2.4
pyobjc-framework-FSEvents==2.4
pyobjc-framework-InputMethodKit==2.4
pyobjc-framework-InstallerPlugins==2.4
pyobjc-framework-InstantMessage==2.4
pyobjc-framework-LatentSemanticMapping==2.4
pyobjc-framework-LaunchServices==2.4
pyobjc-framework-Message==2.4
pyobjc-framework-PreferencePanes==2.4
pyobjc-framework-PubSub==2.4
pyobjc-framework-QTKit==2.4
pyobjc-framework-Quartz==2.4
pyobjc-framework-ScreenSaver==2.4
pyobjc-framework-ScriptingBridge==2.4
pyobjc-framework-SearchKit==2.4
pyobjc-framework-ServerNotification==2.4
pyobjc-framework-ServiceManagement==2.4
pyobjc-framework-Social==2.4
pyobjc-framework-SyncServices==2.4
pyobjc-framework-SystemConfiguration==2.4
pyobjc-framework-WebKit==2.4
BeautifulSoup==3.2.1
requests==0.14.0
